+++ src/mesa/shader/prog_uniform.c
@@ -134,7 +134,7 @@ _mesa_longest_uniform_name(const struct gl_uniform_list *list)
    GLint max = 0;
    GLuint i;
    for (i = 0; list && i < list->NumUniforms; i++) {
-      GLuint len = _mesa_strlen(list->Uniforms[i].Name);
+      GLint len = (GLint)_mesa_strlen(list->Uniforms[i].Name);
       if (len > max)
          max = len;
    }
